+++ ArangoDB.Extensions.VectorData.Tests/IntegrationTests/ArangoCollectionIntegrationTests.EnsureCollectionDeletedAsync.cs
