+++ ArangoDB.Extensions.VectorData.Tests/IntegrationTests/ArangoCollectionIntegrationTests.EnsureCollectionExistsAsync.cs
