+++ ArangoDB.Extensions.VectorData.Tests/IntegrationTests/ArangoCollectionIntegrationTests.GetAsyncWithFilter.EqualityOperator.cs
