+++ ArangoDB.Extensions.VectorData.Tests/IntegrationTests/ArangoCollectionIntegrationTests.cs
