+++ ArangoDB.Extensions.VectorData.Tests/IntegrationTests/ArangoHybridSearchableIntegrationTests.cs
