+++ ArangoDB.Extensions.VectorData.Tests/UnitTests/ArangoCollectionUnitTests.EnsureCollectionDeletedAsync.cs
