+++ ArangoDB.Extensions.VectorData/ArangoCollection.cs
