+++ ArangoDB.Extensions.VectorData/ArangoDB.Extensions.VectorData.csproj
