+++ ArangoDB.Extensions.VectorData/ArangoDynamicCollection.cs
