+++ ArangoDB.Extensions.VectorData/ArangoHybridSearchable.cs
