+++ ArangoDB.Extensions.VectorData/ArangoVectorSearchable.cs
