+++ ArangoDB.Extensions.VectorData/CursorRow.cs
