+++ ArangoDB.Extensions.VectorData/DependencyRegistration.cs
