+++ ArangoDB.Extensions.VectorData/Helpers/AqlFilters.cs
