+++ ArangoDB.Extensions.VectorData/Helpers/AqlLikeWildcardPositions.cs
