+++ ArangoDB.Extensions.VectorData/Helpers/AqlParameterizedQueryHelpers.cs
