+++ ArangoDB.Extensions.VectorData/Helpers/FilterRetrievalOptions.cs
