+++ ArangoDB.Extensions.VectorData/Helpers/KeyHelpers.cs
