+++ ArangoDB.Extensions.VectorData/Helpers/LinqExpressionHelpers/BinaryExpressionHelpers.cs
