+++ ArangoDB.Extensions.VectorData/Helpers/SortOperationHelpers.cs
