+++ ArangoDB.Extensions.VectorData/Helpers/StringComparisonHelpers.cs
