+++ ArangoDB.Extensions.VectorData/Helpers/TypeHelpers.cs
