+++ ArangoDB.Extensions.VectorData/HybridSearchRow.cs
