+++ ArangoDB.KernelMemory/ArangoDB.KernelMemory.csproj
